requirements.txt
dj-database-url==0.3.0
gunicorn==19.4.1
wheel==0.24.0
whitenoise==2.0.6
django==1.9
djangorestframework==3.3.2
psycopg2==2.6.1
django-cors-headers==1.1.0
pillow==3.0.0
dj-static==0.0.6
django-storages==1.1.8
boto==2.38.0
